+++ src/GluttonyCombo/GluttonyCombo/AutoRotation/AutoRotationController.cs
@@ -1581,7 +1581,7 @@ internal unsafe class AutoRotationController
                     if (GetRoleFromJob(Player.Job) is JobRole.MagicalDPS)
                     {
                         if (ActionReady(RoleActions.Magic.Swiftcast) && !HasStatusEffect(RDM.Buffs.Dualcast) &&
-                            !HasFreeInstantCasts && !HasOccultDualcast)
+                            !HasOrExpectsOccultInstantCast)
                         {
                             if (ActionManager.Instance()->GetActionStatus(ActionType.Action, RoleActions.Magic.Swiftcast) == 0)
                             {
@@ -1592,7 +1592,7 @@ internal unsafe class AutoRotationController
                     }
 
                     if (HasStatusEffect(RoleActions.Magic.Buffs.Swiftcast) || HasStatusEffect(RDM.Buffs.Dualcast) ||
-                        HasFreeInstantCasts || HasOccultDualcast || !IsMoving())
+                        HasOrExpectsOccultInstantCast || !IsMoving())
                     {
                         ActionManager.Instance()->UseAction(ActionType.Action, resSpell, member.BattleChara.GameObjectId);
                         return true;
@@ -1602,7 +1602,7 @@ internal unsafe class AutoRotationController
                 if (Player.Job is Job.RDM)
                 {
                     if (ActionReady(RoleActions.Magic.Swiftcast) && !HasStatusEffect(RDM.Buffs.Dualcast) &&
-                        !HasFreeInstantCasts && !HasOccultDualcast)
+                        !HasOrExpectsOccultInstantCast)
                     {
                         ActionManager.Instance()->UseAction(ActionType.Action, RoleActions.Magic.Swiftcast);
                         return true;
@@ -1618,7 +1618,7 @@ internal unsafe class AutoRotationController
                 else
                 {
                     if (ActionReady(RoleActions.Magic.Swiftcast) &&
-                        !HasFreeInstantCasts && !HasOccultDualcast)
+                        !HasOrExpectsOccultInstantCast)
                     {
                         if (ActionManager.Instance()->GetActionStatus(ActionType.Action, RoleActions.Magic.Swiftcast) == 0)
                         {
@@ -1628,7 +1628,7 @@ internal unsafe class AutoRotationController
                     }
 
                     if (!IsMoving() || HasStatusEffect(RoleActions.Magic.Buffs.Swiftcast) ||
-                        HasFreeInstantCasts || HasOccultDualcast)
+                        HasOrExpectsOccultInstantCast)
                     {
 
                         if ((cfg is not null) && ((cfg.HealerSettings.AutoRezRequireSwift && ActionManager.GetAdjustedCastTime(ActionType.Action, resSpell) == 0) || !cfg.HealerSettings.AutoRezRequireSwift))